Added mockNimServingRuntimeTemplate mock function
